Bind threads to an unused CPU on a specific numa node
The version here is rudimentary and assumes that each DAG
stream is alternately bound to NUMA nodes 0 and 1.

It should be possible to use the dagconfig API to find out
the right NUMA node beforehand rather than assuming, but that's
a job for another day.

Add check for libnuma to configure.ac