Remove unused code from libndag-server
Push record API function is now gone.
Compression code is now gone -- had a terrible impact on
performance, but we can maybe revisit this another time.

Removed +1's from iovec indexing in push iovec API -- this
was a holdover from an early implementation that had assumed
we wanted to reserve an iovec for the ndag header but this
is not quite how it worked out.