Extend stats logging to support logging to a file
Each stream writes to a separate file within a given logging directory that is overwritten each time stats are dumped. This way an external process can watch these files and extract stats. This saves dealing with rotation etc. and makes it very hard for the stats subsystem to crash the multicaster itself (as could happen if we used some kind of fancier message queue or some such).