Allow darkfilter exclusion list to be reloaded
We now share the filter state among threads and keep a flip-flop buffer for the exclusion array. When a reload is requested, the not-in-use buffer is reset and loaded with the new exclusion list and then the "current_exlude" value is flipped to send the processing threads to this exclude array. If updating the exclusion list fails, the updater thread emits a warning, but keeps running, using the previous (good) exlude array.