Attach DAG streams only in the thread that will be using them
We previously had to attach the stream in the main thread to check
if it was valid, and figured we could re-use this attachment in
our per-stream threads. This is not the case: you end up with all
sorts of concurrency issues which manifest as bits of packets from
other streams suddenly appearing in the wrong stream, packet
corruption etc.

Now we do our initial attach to check validity, detach, and then
attach again inside the per-stream thread.