Wrap all dag stream attachment and starting operations in a mutex.
Turns out these functions may not be entirely thread-safe, so
it's safer to make sure only one thread is calling them at
any given time.

Helgrind also gets somewhat angry at calls to dag_advance_stream()
but it seems relatively innocuous so far and not worth the
performance hit of grabbing a mutex everytime we want to advance
the DAG read pointer. May need to re-evaluate if we see crashes
or corruption though...